vasanth-chatbot/requirements.txt
google-generativeai>=0.8.4
flask==2.3.3
flask-socketio==5.3.4
python-dotenv==1.0.0
eventlet==0.33.3